:recycle: Restructure the directory layout of the CLI wrapper
