+++ cli/main.node.ts
@@ -5,7 +5,7 @@
  */
 
 import { dirname } from "jsr:@std/path@1.1.4";
-import * as git from "../git-wrapper/mod.ts";
+import * as git from "../wrapper/git/mod.ts";
 import { ghDescribeCli } from "./cli.ts";
 
 declare const __filename: string;
