+++ cli/main.ts
@@ -1,5 +1,5 @@
 import { dirname, fromFileUrl } from "jsr:@std/path@1.1.4";
-import * as git from "../git-wrapper/mod.ts";
+import * as git from "../wrapper/git/mod.ts";
 import { ghDescribeCli } from "./cli.ts";
 
 async function version(): Promise<string> {
